// touristagency-front3/generated/models/CarMark.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tourist agency
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CarMark
 */
export interface CarMark {
    /**
     * 
     * @type {string}
     * @memberof CarMark
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof CarMark
     */
    id: string;
}

/**
 * Check if a given object implements the CarMark interface.
 */
export function instanceOfCarMark(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;
    isInstance = isInstance && "id" in value;

    return isInstance;
}

export function CarMarkFromJSON(json: any): CarMark {
    return CarMarkFromJSONTyped(json, false);
}

export function CarMarkFromJSONTyped(json: any, ignoreDiscriminator: boolean): CarMark {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'id': json['id'],
    };
}

export function CarMarkToJSON(value?: CarMark | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'id': value.id,
    };
}